Add categories to the admin panel; Clean up the admin panel

+++ src/main/java/org/jastka4/pwr/idb/repository/CategoryRepository.java
@@ -4,5 +4,5 @@ import org.jastka4.pwr.idb.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
-    Category findById(final int id);
+    Category findById(final long id);
 }
